decoupling repository from api repository
