+++ app/src/main/java/com/hanitacm/weatherapp/domain/GetWeatherUseCase.kt
@@ -1,11 +1,12 @@
 package com.hanitacm.weatherapp.domain
 
 import com.hanitacm.weatherapp.repository.WeatherRepository
+import com.hanitacm.weatherapp.repository.api.WeatherApi
 import io.reactivex.Single
 
-class GetWeatherUseCase(weatherRepository: WeatherRepository) {
+class GetWeatherUseCase(val weatherRepository: WeatherRepository) {
   fun getWeather(location: String): Single<String> {
-    return Single.just<String>("")
+    return weatherRepository.getWeather(location)
   }
 
 }
