+++ app/src/main/java/com/hanitacm/weatherapp/domain/WeatherDomainModel.kt
@@ -1,0 +1,5 @@
+package com.hanitacm.weatherapp.domain
+
+class WeatherDomainModel {
+
+}
